feat(worker): merge vsl staking and bridge to vsl core (#459)
