+++ internal/database/driver.go
@@ -3,7 +3,6 @@ package database
 type Driver string
 
 const (
-	DriverCockroachDB Driver = "cockroachdb"
-	DriverPostgreSQL  Driver = "postgresql"
-	DriverMySQL       Driver = "mysql"
+	DriverPostgreSQL Driver = "postgresql"
+	DriverMySQL      Driver = "mysql"
 )
