+++ schema/worker/decentralized/worker.go
@@ -32,7 +32,6 @@ const (
 	SAVM                         // savm
 	Stargate                     // stargate
 	Uniswap                      // uniswap
-	VSL                          // vsl
 )
 
 func (w Worker) Component() string {
@@ -84,5 +83,4 @@ var ToTagsMap = map[Worker][]tag.Tag{
 	SAVM:       {tag.Transaction},
 	Stargate:   {tag.Transaction},
 	Uniswap:    {tag.Exchange, tag.Transaction},
-	VSL:        {tag.Transaction},
 }
